LayoutTests/storage/indexeddb/object-lookups-in-versionchange.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
<script>

description("Regression test for http://webkit.org/b/102547");

indexedDBTest(prepareDatabase, finishJSTest);

function prepareDatabase() {
    evalAndLog("db = event.target.result");
    evalAndLog("transaction = event.target.transaction");
    evalAndLog("store = db.createObjectStore('store')");
    debug("");
    evalAndExpectException("transaction.objectStore('no-such-store')", "DOMException.NOT_FOUND_ERR", "'NotFoundError'");
    debug("");
    evalAndExpectException("store.index('no-such-index')", "DOMException.NOT_FOUND_ERR", "'NotFoundError'");
}

</script>
<script src=../../resources/js-test-post.js></script>
